RESTful refactoring of /course access continued Move index access into the url Move course creation into the url Add helper methods for testing to serialize json data and set accept header.
