+++ cms/djangoapps/contentstore/tests/utils.py
@@ -30,12 +30,24 @@ def registration(email):
 
 
 class AjaxEnabledTestClient(Client):
+    """
+    Convenience class to make testing easier.
+    """
     def ajax_post(self, path, data=None, content_type="application/json", **kwargs):
+        """
+        Convenience method for client post which serializes the data into json and sets the accept type
+        to json
+        """
         if not isinstance(data, basestring):
             data = json.dumps(data or {})
         kwargs.setdefault("HTTP_X_REQUESTED_WITH", "XMLHttpRequest")
         return self.post(path=path, data=data, content_type=content_type, **kwargs)
 
+    def get_html(self, path, data=None, follow=False, **extra):
+        """
+        Convenience method for client.get which sets the accept type to html
+        """
+        return self.get(path, data or {}, follow, HTTP_ACCEPT="text/html", **extra)
 
 @override_settings(MODULESTORE=TEST_MODULESTORE)
 class CourseTestCase(ModuleStoreTestCase):
