+++ common/djangoapps/student/tests/test_userstanding.py
@@ -53,7 +53,7 @@ class UserStandingTest(TestCase):
         try:
             self.some_url = reverse('dashboard')
         except NoReverseMatch:
-            self.some_url = reverse('index')
+            self.some_url = '/course'
 
         # since it's only possible to disable accounts from lms, we're going
         # to skip tests for cms
